tests/wpt/mozilla/tests/mozilla/bluetooth/getPrimaryServices/get-same-service.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.two_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => Promise.all([
        gattServer.getPrimaryServices(heart_rate.name),
        gattServer.getPrimaryServices(heart_rate.name)]))
    .then(services => {
        let services1 = services[0];
        let services2 = services[1];
        assert_equals(services1.length, services2.length);
        for (let i = 0; i < services1.length; i++) {
            assert_equals(services1[i], services2[i],
            'Should return the same service as the first call.');
        }
    });
}, 'Calls to get the same service should return the same object.');
</script>
